requirements.txt
asgiref
beautifulsoup4
celery
emoji
numpy
Django
djangorestframework==3.14.0
django-allauth==0.54.0
django-auditlog==2.3.0
django-snowflake
django-celery-beat
django-snowflake
django-cors-headers
dj-rest-auth
django-softdelete
django-role-permissions
djangorestframework-simplejwt
djangorestframework-jwt==1.11.0
drf-yasg
google-api-python-client
google-cloud-dialogflow-cx
google-auth
google-auth-oauthlib
gunicorn
instagrapi
mailchimp-marketing
openpyxl
pandas
Pillow
pre-commit
PyJWT
pytz
psycopg2-binary
redis
requests
selenium
sentry-sdk[django]
scrapy
sqlparse
coverage
robotframework
webdriver-manager
tabulate



